push temp

+++ temp/edit_char_add.c
@@ -6,7 +6,7 @@
 /*   By: cobrecht <cobrecht@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/02/16 16:17:05 by cobrecht          #+#    #+#             */
-/*   Updated: 2014/02/16 17:11:27 by cobrecht         ###   ########.fr       */
+/*   Updated: 2014/02/16 18:30:15 by cobrecht         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 static t_char	*to_front(t_char **nchr, t_char **lst, t_cmd *cmd, int *c_pos);
 static void		to_back(t_char **newchar, t_char **list);
 
-t_char			*edit_char_add(t_char *list, int chr, int *c_pos, t_cmd *cmd)
+t_char			*edit_char_add(t_char *list, long chr, int *c_pos, t_cmd *cmd)
 {
 	t_char		*newchar;
 
 	newchar = NULL;
 	if (!(newchar = (t_char *)malloc(sizeof(t_char))))
 		return (NULL);
-	newchar->c = chr;
+	newchar->c = (int)chr;
 	newchar->next = NULL;
 	if (*c_pos == 0)
 	{
